[STEELAI-869] Fix mfd, misc, and ptp compilation issues
Includes fix for:

  https://lore.kernel.org/oe-kbuild-all/202501110703.AqVudeRO-lkp@intel.com/
    drivers/misc/rsmu_cm.c:172:13: warning: variable 'err' set but not used

  https://lore.kernel.org/oe-kbuild-all/202501110745.sHP4ZB8s-lkp@intel.com/
    drivers/misc/rsmu_cm.c:333:53: warning: implicit conversion from 'int' to 'u16' (aka 'unsigned short') changes value

+++ linux/drivers/mfd/rsmu_spi.c
@@ -226,7 +226,7 @@ static int rsmu_spi_probe(struct spi_device *client)
 
 	rsmu->regmap = devm_regmap_init(&client->dev, NULL, client, cfg);
 	if (IS_ERR(rsmu->regmap)) {
-		ret = PTR_ERR(rsmu->regmap);
+		ret = (int)PTR_ERR(rsmu->regmap);
 		dev_err(rsmu->dev, "Failed to allocate register map: %d\n", ret);
 		return ret;
 	}
